update given, when then steps in mocha specs
